refactor archive ops into a separate file
